rename and move services to service folder
