+++ src/client/app/test/__tests__/components/routes/cart.test.tsx
@@ -5,8 +5,8 @@ import { CartContext } from '../../../../store/cart-context';
 import { mockCartContextValue } from '../../../__mocks__/store/cart-context-mock';
 import { type renderCartContextProps } from '../../../__mocks__/store/context-mock';
 
-jest.mock('../../../../util/fetch-carts-products', () => ({
-  fetchCartsProducts: jest.fn(() => Promise.resolve([])),
+jest.mock('../../../../service/carts-products-service', () => ({
+  getCartsProducts: jest.fn(() => Promise.resolve([])),
 }));
 
 let container: HTMLElement;
